// src/generated/speedy/types.gen.ts
export type ShipmentParcelSize = {
    /**
     * Parcel width in cm.
     */
    width: number;
    /**
     * Parcel depth in cm.
     */
    depth: number;
    /**
     * Parcel height in cm.
     */
    height: number;
};
// ...
export type ShipmentAddress = {
    /**
     * Country ISO code. If not provided, local country is assumed. Used for all address types. Constraints: Validate for valid country code.
     */
    countryId?: number;
    /**
     * Country state. Used for addresses of type 2 (foreign address). Constraints: Validate for valid country state.
     */
    stateId?: string;
    /**
     * Site id. Used for all address types. Constraints: Validate for valid site and value is required.
     */
    siteId?: number;
    /**
     * Site type can be used in conjunction with countryId and siteName to find unique site. Used for addresses of type 1 (local address). Constraints: Max 20
     */
    siteType?: string;
    /**
     * Site type can be used in conjunction with countryId and siteName to find unique site. Used for all address types. Constraints: Max 50
     */
    siteName?: string;
    /**
     * Can be used in conjunction with countryId to find unique site. Used for all address types. Constraints: Validated for valid postcode in site and country. Max 10
     */
    postCode?: string;
    /**
     * Street identifier. If not provided, but street type and street name are provided - the system will try to find unique match by them in site.
     * Used for addresses of type 1 (local address). Constraints: Validate for valid street.
     */
    streetId?: number;
    /**
     * Street type. Used for addresses of type 1 (local address). Constraints: Max 15
     */
    streetType?: string;
    /**
     * Street name. Used for addresses of type 1 (local address). Constraints: Max 50
     */
    streetName?: string;
    /**
     * Street number. Used for addresses of type 1 (local address). Constraints: Max 10
     */
    streetNo?: string;
    /**
     * Complex identifier. If not provided, but complex type and complex name are provided - the system will try to find unique match by them in site.
     * Used for addresses of type 1 (local address). Constraints: Validate for valid complex.
     */
    complexId?: number;
    /**
     * Complex type. Used for addresses of type 1 (local address). Constraints: Max 15
     */
    complexType?: string;
    /**
     * Complex name. Used for addresses of type 1 (local address). Constraints: Max 50
     */
    complexName?: string;
    /**
     * Block No. Used for addresses of type 1 (local address). Constraints: Max 32
     */
    blockNo?: string;
    /**
     * Entrance. Used for addresses of type 1 (local address). Constraints: Max 10
     */
    entranceNo?: string;
    /**
     * Floor. Used for addresses of type 1 (local address). Constraints: Max 10
     */
    floorNo?: string;
    /**
     * Apartment. Used for addresses of type 1 (local address). Constraints: Max 10
     */
    apartmentNo?: string;
    /**
     * Point of interest identifier. Used for addresses of type 1 (local address). Constraints: Validate for valid point of interest.
     */
    poiId?: number;
    /**
     * Address note. Used for addresses of type 1 (local address). Constraints: 200
     */
    addressNote?: string;
    /**
     * Address line 1. Used for addresses of type 2 (foreign address). Constraints: Max 35
     */
    addressLine1?: string;
    /**
     * Address line 2. Used for addresses of type 2 (foreign address). Constraints: Max 35
     */
    addressLine2?: string;
    /**
     * GIS coordinates - X .Used for all address types.
     */
    x?: number;
    /**
     * GIS coordinates - Y. Used for all address types.
     */
    y?: number;
};

export type Address = ShipmentAddress & {
    /**
     * Full address in a single text field
     */
    fullAddressString: string;
    /**
     * The address site description in a single text
     */
    siteAddressString: string;
    /**
     * The address within the site in a single text
     */
    localAddressString: string;
};
// ...
export type Office = {
    /**
     * Office id
     */
    id?: number;
    /**
     * Office name in requested language
     */
    name?: string;
    /**
     * Office name in English
     */
    nameEn?: string;
    /**
     * Office site id
     */
    siteId?: number;
    /**
     * Office address in requested language
     */
    address?: Address;
    /**
     * Office work time start for days with standard (FULL) working schedule
     */
    workingTimeFrom?: string;
    /**
     * Office work time end for days with standard (FULL) working schedule
     */
    workingTimeTo?: string;
    /**
     * Office work time start for days with saturday (HALF) working schedule
     */
    workingTimeHalfFrom?: string;
    /**
     * Office work time end for days with saturday (HALF) working schedule
     */
    workingTimeHalfTo?: string;
    /**
     * Office work time start for non-working days (DAY OFF)
     */
    workingTimeDayOffFrom?: string;
    /**
     * Office work time end for non-working days (DAY OFF)
     */
    workingTimeDayOffTo?: string;
    /**
     * Office same day departure cutoff for days with standard (FULL) working schedule (departure of parcels delivered to office after the cutoff time is next working day)
     */
    sameDayDepartureCutoff?: string;
    /**
     * Office same day departure cutoff for days with saturday (HALF) working schedule (departure of parcels delivered to office after the cutoff time is next working day)
     */
    sameDayDepartureCutoffHalf?: string;
    /**
     * Office same day departure cutoff for non-working days (DAY OFF) working schedule (departure of parcels delivered to office after the cutoff time is next working day)
     */
    sameDayDepartureCutoffDayOff?: string;
    /**
     * Max parcels dimensions for this office
     */
    maxParcelDimensions?: ShipmentParcelSize;
    /**
     * Max parcel weight
     */
    maxParcelWeight?: number;
    /**
     * Office type
     */
    type?: 'OFFICE' | 'APT';
    /**
     * Nearby office id
     */
    nearbyOfficeId?: number;
    /**
     * Office work time schedule for next days
     */
    workingTimeSchedule?: Array<OfficeWorkingTimeSchedule>;
    /**
     * Pallet office flag
     */
    palletOffice?: boolean;
    /**
     * Flag, indicating this office allows card payments
     */
    cardPaymentAllowed?: boolean;
    /**
     * Flag, indicating this office allows cash payments
     */
    cashPaymentAllowed?: boolean;
    /**
     * Valid from date
     */
    validFrom?: string;
    /**
     * Valid to date
     */
    validTo?: string;
    /**
     * Allowed cargo types
     */
    cargoTypesAllowed?: Array<'PARCEL' | 'PALLET' | 'TYRE'>;
    /**
     * Flag, indicating whether parcels can be picked up from office
     */
    pickUpAllowed?: boolean;
    /**
     * Flag, indicating whether parcels can be dropped-off in office
     */
    dropOffAllowed?: boolean;
    routingInformation?: RoutingInformation;
};

export type OfficeWorkingTimeSchedule = {
    /**
     * Date this schedule is applicable for
     */
    date?: string;
    /**
     * Office working time start for this date
     */
    workingTimeFrom?: string;
    /**
     * Office working time end for this date
     */
    workingTimeTo?: string;
    /**
     * Office same day departure cutoff for this day (departure of parcels delivered to office after the cutoff time is next working day)
     */
    sameDayDepartureCutoff?: string;
    /**
     * If true - it is from standard schedule, otherwise office working time is overridden for this specific day
     */
    standardSchedule?: boolean;
};
// ...
export type RoutingInformation = {
    tour?: Tour;
    /**
     * Destination office id
     */
    officeId?: number;
    /**
     * Destination hub id
     */
    hubId?: number;
    /**
     * Priority
     */
    priority?: number;
};

export type Tour = {
    /**
     * Tour number
     */
    number?: number;
    /**
     * Tour office id
     */
    officeId?: number;
    /**
     * Tour ramp
     */
    ramp?: string;
    /**
     * Tour cell
     */
    cell?: string;
};